+++ .eslintrc
@@ -8,7 +8,7 @@
     "react": { "version": "detect" }
   },
   "env": {
-    "mocha": true
+    "jest": true
   },
   "rules": {
     "react/prop-types": "off",
