+++ package.json
@@ -18,21 +18,17 @@
     "test"
   ],
   "scripts": {
-    "build": "webpack",
-    "build:prod": "NODE_ENV=production npm run build",
-    "compile": "babel src --out-dir lib --copy-files",
-    "compile:prod": "NODE_ENV=production npm run compile",
+    "analyze": "webpack --profile --json | webpack-bundle-size-analyzer",
+    "build": "npm run compile && npm run bundle",
+    "bundle": "webpack --output-pathinfo --progress --hide-modules && NODE_ENV=production webpack --progress --hide-modules",
+    "compile": "babel src --out-dir lib",
     "coveralls": "cat ./coverage/lcov.info | coveralls",
     "dev": "webpack-dev-server --open",
     "format": "prettier --write '{src,test}/**/*.js' && eslint --fix '{src,test}/**/*.js'",
     "lint": "prettier --check '{src,test}/**/*.js' && eslint '{src,test}/**/*.js'",
     "pretest": "npm run lint",
-    "profile": "webpack --profile --json | webpack-bundle-size-analyzer",
-    "profile:prod": "NODE_ENV=production npm run profile",
     "readme": "doctoc --maxlevel 3 README.md",
-    "test": "npm run test:node && npm run test:browser",
-    "test:node": "NODE_ENV=testing nyc mocha",
-    "test:browser": "NODE_ENV=testing karma start",
+    "test": "jest --coverage && JEST_ENV=jsdom jest",
     "update": "ncu -u"
   },
   "dependencies": {
@@ -58,7 +54,6 @@
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.0.0-beta.2",
     "babel-plugin-istanbul": "^4.1.6",
-    "chai": "^4.1.2",
     "coveralls": "^3.0.0",
     "css-loader": "^0.28.10",
     "doctoc": "^1.4.0",
@@ -76,20 +71,11 @@
     "eslint-plugin-standard": "^4.0.1",
     "file-loader": "^1.1.11",
     "husky": "^3.0.8",
-    "jsdom": "^15.1.1",
-    "karma": "^2.0.0",
-    "karma-jsdom-launcher": "^6.1.3",
-    "karma-mocha": "^1.3.0",
-    "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "^2.0.13",
-    "mocha": "^5.0.4",
+    "jest": "^25.2.4",
     "npm-check-updates": "^4.1.2",
-    "nyc": "^14.1.1",
     "prettier": "^2.0.2",
     "react-hot-loader": "^4.0.0",
     "referencer": "^0.2.5",
-    "regenerator-runtime": "^0.11.1",
-    "sinon": "^4.4.4",
     "style-loader": "^0.20.3",
     "webpack": "^4.1.1",
     "webpack-bundle-size-analyzer": "^2.7.0",
