--- test/setup.js
@@ -1,17 +1,0 @@
-// Helpers
-
-async function catchError(func, ...args) {
-  let error
-  try {
-    await func(...args)
-  } catch (exception) {
-    error = exception
-  }
-  return error
-}
-
-// System
-
-module.exports = {
-  catchError,
-}
